# docs/spec/api-v1-map.openapi.json
{
  "components": {
    "schemas": {
      "EnsembleResponse": {
        "properties": {
          "forecast_horizon_steps": {
            "title": "Forecast Horizon Steps",
            "type": "integer"
          },
          "member_count": {
            "title": "Member Count",
            "type": "integer"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "representation": {
            "title": "Representation",
            "type": "string"
          },
          "series": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Series",
            "type": "object"
          },
          "time_step_seconds": {
            "title": "Time Step Seconds",
            "type": "integer"
          },
          "units": {
            "title": "Units",
            "type": "string"
          },
          "valid_times": {
            "items": {
              "format": "date-time",
              "type": "string"
            },
            "title": "Valid Times",
            "type": "array"
          }
        },
        "required": [
          "representation",
          "parameter",
          "units",
          "forecast_horizon_steps",
          "time_step_seconds",
          "member_count",
          "valid_times",
          "series"
        ],
        "title": "EnsembleResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ForecastDetail": {
        "properties": {
          "combination_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combination Strategy"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ensemble": {
            "$ref": "#/components/schemas/EnsembleResponse"
          },
          "forecast_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Status"
          },
          "forecast_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Version"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Quality"
          },
          "input_quality_flags": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Quality Flags"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "model_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Artifact Id"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_tier": {
            "title": "Model Tier",
            "type": "string"
          },
          "nwp_cycle_reference_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nwp Cycle Reference Time"
          },
          "nwp_cycle_source": {
            "title": "Nwp Cycle Source",
            "type": "string"
          },
          "observation_staleness_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Staleness Hours"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "preservation_at_publish": {
            "anyOf": [
              {
                "enum": [
                  "verified",
                  "backup_pending"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preservation At Publish"
          },
          "publication_decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Decision Id"
          },
          "publication_state": {
            "anyOf": [
              {
                "enum": [
                  "unpublished",
                  "selected",
                  "replaced",
                  "withdrawn"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication State"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "qc_flags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/QcFlagResponse"
            },
            "title": "Qc Flags",
            "type": "array"
          },
          "qc_status": {
            "enum": [
              "raw",
              "qc_passed",
              "qc_suspect",
              "qc_failed",
              "missing",
              "qc_unchecked"
            ],
            "title": "Qc Status",
            "type": "string"
          },
          "representation": {
            "title": "Representation",
            "type": "string"
          },
          "selection_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Version"
          },
          "source_model_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Model Ids"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "warm_up_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warm Up Source"
          }
        },
        "required": [
          "id",
          "station_id",
          "model_id",
          "model_tier",
          "issued_at",
          "parameter",
          "representation",
          "status",
          "qc_status",
          "nwp_cycle_source",
          "created_at",
          "version",
          "updated_at",
          "ensemble"
        ],
        "title": "ForecastDetail",
        "type": "object"
      },
...
      "ForecastSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "forecast_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Status"
          },
          "forecast_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Version"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Quality"
          },
          "input_quality_flags": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Quality Flags"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_tier": {
            "title": "Model Tier",
            "type": "string"
          },
          "nwp_cycle_source": {
            "title": "Nwp Cycle Source",
            "type": "string"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "preservation_at_publish": {
            "anyOf": [
              {
                "enum": [
                  "verified",
                  "backup_pending"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preservation At Publish"
          },
          "publication_decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Decision Id"
          },
          "publication_state": {
            "anyOf": [
              {
                "enum": [
                  "unpublished",
                  "selected",
                  "replaced",
                  "withdrawn"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication State"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "qc_flags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/QcFlagResponse"
            },
            "title": "Qc Flags",
            "type": "array"
          },
          "qc_status": {
            "enum": [
              "raw",
              "qc_passed",
              "qc_suspect",
              "qc_failed",
              "missing",
              "qc_unchecked"
            ],
            "title": "Qc Status",
            "type": "string"
          },
          "representation": {
            "title": "Representation",
            "type": "string"
          },
          "selection_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Version"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "station_id",
          "model_id",
          "model_tier",
          "issued_at",
          "parameter",
          "representation",
          "status",
          "qc_status",
          "nwp_cycle_source",
          "created_at"
        ],
        "title": "ForecastSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse_ForecastSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ForecastSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[ForecastSummary]",
        "type": "object"
      },
...
      "PublicationTombstone": {
        "properties": {
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "forecast_id": {
            "title": "Forecast Id",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "publication_state": {
            "const": "withdrawn",
            "default": "withdrawn",
            "title": "Publication State",
            "type": "string"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "withdrawn_at": {
            "format": "date-time",
            "title": "Withdrawn At",
            "type": "string"
          }
        },
        "required": [
          "forecast_id",
          "station_id",
          "parameter",
          "issued_at",
          "decision_id",
          "withdrawn_at"
        ],
        "title": "PublicationTombstone",
        "type": "object"
      },
      "QcFlagResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "rule_version": {
            "title": "Rule Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "qc_passed",
              "qc_suspect",
              "qc_failed",
              "qc_unchecked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "rule_version",
          "status"
        ],
        "title": "QcFlagResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The subset of /api/v1 the flow map reads (Plan 402). Not served \u2014 openapi_url stays None on the live app; this is a committed, drift-tested static artefact. Bearer auth is added explicitly below: the live routes read the Authorization header from the raw Request, so FastAPI's own generator sees no security scheme.",
    "title": "SAPPHIRE Flow \u2014 flow map contract",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/forecasts/{forecast_id}": {
      "get": {
        "description": "Forecast detail, with typed qc_flags. For an activated CHWRR tenant only selected or historical replaced publications expose values; an unpublished or out-of-scope ID is 404, a withdrawn ID is 410 with a metadata-only tombstone. All service roles use the same publication gate. A malformed forecast_id remains unguarded and returns 500, not 400.",
        "operationId": "get_forecast_api_v1_forecasts__forecast_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "forecast_id",
            "required": true,
            "schema": {
              "title": "Forecast Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown or out-of-scope station/forecast \u2014 the same body either way (never reveals which)."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationTombstone"
                }
              }
            },
            "description": "Withdrawn forecast; metadata-only tombstone."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Forecast",
        "tags": [
          "api-forecasts"
        ]
      }
    },
...
    "/api/v1/stations/{station_id}/forecasts": {
      "get": {
        "description": "Forecast summaries for a station, with QC status and typed qc_flags. For an activated CHWRR tenant this contains only the current human-selected publication per issue time, including a selected forecast later marked superseded by an automatic retry. Consumer and reviewer scopes are identical; admin is unscoped but sees the same publication gate. An out-of-scope station is 404.",
        "operationId": "list_forecasts_api_v1_stations__station_id__forecasts_get",
        "parameters": [
          {
            "in": "path",
            "name": "station_id",
            "required": true,
            "schema": {
              "title": "Station Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "in": "query",
            "name": "parameter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameter"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "description": "Return only forecasts assessed PARTIAL or DEGRADED input quality (Plan 253 OD-2). Forecasts with no assessment recorded (unknown) are never included.",
            "in": "query",
            "name": "degraded_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "Return only forecasts assessed PARTIAL or DEGRADED input quality (Plan 253 OD-2). Forecasts with no assessment recorded (unknown) are never included.",
              "title": "Degraded Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ForecastSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed query value or station_id (error envelope: {\"error\": <message>, \"detail\": null})."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown or out-of-scope station/forecast \u2014 the same body either way (never reveals which)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Forecasts",
        "tags": [
          "api-stations"
        ]
      }
    },
...
  }
}